solved homework #9

+++ les9/include/mathFunc.h
@@ -5,6 +5,6 @@ float sum   (float a, float b);
 float sub   (float a, float b);
 float divide(float a, float b);
 float mul   (float a, float b);
-float op_log(int   base, float val);
+float op_log(float base, float val);
 
 #endif //__MATH_FUNCTIONS__
